requirements.txt
opencv-python==4.7.0.68
numpy==1.23.4
scipy==1.9.3
networkx==3.0
matplotlib==3.6.3
pyyaml==6.0
tqdm==4.64.1
torch==1.13.1
tensorboard==2.11.2
cvxpy==1.3.0
seaborn==0.12.2
torch_geometric==2.3.0
pybullet==3.2.5
pillow==9.5.0